Read TODOs from today with todo.sh
Insert done TODOs from the day of writing the day into the template.

This patch relies on usage of the todo.sh utility for managing TODOs.
The way the TODOs are extracted and cleaned up is very specific to how
my todo.sh is configured and what add-ons I use.

There might be a way to make this more more generic, making it easier
for others to use this script, but this is currently not a concern of
mine.