Remove "Tags" section from template
I added the "Tags" section as a result of being inspired by a diary app
on my phone (Daylio): tags were the first thing one entered, completing
the entry with some free text summary.

Since the current version of the script enters TODOs from the day,
there's already a summary of what transpired, making tags on top of that
unnecessary. On top of that, I prefer to write at length in my diary,
instead of resorting to short entries like I did on my phone.