Show previous entry in pager
Not only do I find it nicer to read entries in a pager, it'll also work
better when I'll start concatenating entries.

While I was considering how one could avoid writing a file for the pager
to read, I decided that it was easier to just use Tempfile and be done
with it.

Usage of the ENV variable "DIARY_PAGER" instead of "PAGER" is
deliberate: I'm currently using "nvim" with a bunch of options as my
DIARY_PAGER (mainly to make use of Goyo.vim), which is quite different
from my regular PAGER. If these two would at some point be the same,
then I'll just set DIARY_PAGER to use the value of PAGER.