Show full day instead of just random entry
I know I'll be adding multiple entries during the same day, from time to
time. Currently, if I was to add a very short entry and the random
selection of past entry would select this short entry, it would probably
not be a very meaningful retrospective. Instead, pick a day randomly out
of all the days with entries and concatenate all the entries from that
day.

This commit introduces a fair bit of logic and we're rapidly approaching
the point of refactoring time.