Entries should not contain done TODOs
Since the TODOs for a specific day can be obtained by querying the
`todo.sh` utility, there's no need to include this text into the final
entry files. Having the TODOs present for viewing when writing the entry
is useful to remember what has happened during the day, but the this
text should be removed when saving the entry.

Tweak the `delete_and_exit_if_empty_entry` method to just check whether
there are only lines with whitespace left, instead of checking against
the contents of the original template.

Generalize the `Todos#done_today` method to `Todos#done_for`, so a date
can be provided as a parameter.